+++ CHANGELOG.md
@@ -24,6 +24,13 @@ following conventions:
 
 [KaC]: <https://keepachangelog.com/en/1.0.0/>
 
+## 0.3.0.0 (2020-11-03)
+
+### Breaking
+
+* Use `Textual` error messages
+* Add `maybeParseWithRead` function
+
 ## 0.2.3.0 (2020-09-25)
 
 ### Non-Breaking